feat：申请加入用户组 itsm单据内容不详细问题 #9365
