+++ src/backend/ci/core/auth/biz-auth-rbac/src/main/kotlin/com/tencent/devops/auth/service/PermissionGradeManagerService.kt
@@ -175,7 +175,7 @@ class PermissionGradeManagerService @Autowired constructor(
             gradeManagerId
         } else {
             val callbackId = UUIDUtil.generate()
-            val itsmContentDTO = buildItsmContentDTO(
+            val itsmContentDTO = itsmService.buildGradeManagerItsmContentDTO(
                 projectName = projectName,
                 projectId = projectCode,
                 desc = projectApprovalInfo.description ?: "",
@@ -290,7 +290,7 @@ class PermissionGradeManagerService @Autowired constructor(
             true
         } else {
             val callbackId = UUIDUtil.generate()
-            val itsmContentDTO = buildItsmContentDTO(
+            val itsmContentDTO = itsmService.buildGradeManagerItsmContentDTO(
                 projectName = projectName,
                 projectId = projectCode,
                 desc = projectApprovalInfo.description ?: "",
@@ -616,48 +616,5 @@ class PermissionGradeManagerService @Autowired constructor(
                 resourceName = projectName
             )
         )
-    }
-
-    @Suppress("LongParameterList")
-    private fun buildItsmContentDTO(
-        projectName: String,
-        projectId: String,
-        desc: String,
-        organization: String,
-        authSecrecy: Int,
-        subjectScopes: List<SubjectScopeInfo>
-    ): ItsmContentDTO {
-        val itsmColumns = listOf(
-            ItsmColumn.builder().key("projectName")
-                .name(I18nUtil.getCodeLanMessage(BK_PROJECT_NAME)).type("text").build(),
-            ItsmColumn.builder().key("projectId").name(I18nUtil.getCodeLanMessage(BK_PROJECT_ID)).type("text").build(),
-            ItsmColumn.builder().key("desc").name(I18nUtil.getCodeLanMessage(BK_PROJECT_DESC)).type("text").build(),
-            ItsmColumn.builder().key("organization")
-                .name(I18nUtil.getCodeLanMessage(BK_ORGANIZATION)).type("text").build(),
-            ItsmColumn.builder().key("authSecrecy")
-                .name(I18nUtil.getCodeLanMessage(BK_AUTH_SECRECY)).type("text").build(),
-            ItsmColumn.builder().key("subjectScopes")
-                .name(I18nUtil.getCodeLanMessage(BK_SUBJECT_SCOPES)).type("text").build()
-        )
-        val itsmAttrs = ItsmAttrs.builder().column(itsmColumns).build()
-        val itsmScheme = ItsmScheme.builder().attrs(itsmAttrs).type("table").build()
-        val scheme = HashMap<String, ItsmScheme>()
-        scheme["content_table"] = itsmScheme
-        val value = HashMap<String, ItsmStyle>()
-        value["projectName"] = ItsmStyle.builder().value(projectName).build()
-        value["projectId"] = ItsmStyle.builder().value(projectId).build()
-        value["desc"] = ItsmStyle.builder().value(desc).build()
-        value["organization"] = ItsmStyle.builder().value(organization).build()
-        value["authSecrecy"] =
-            ItsmStyle.builder().value(ProjectAuthSecrecyStatus.getStatus(authSecrecy)?.desc ?: "").build()
-        value["subjectScopes"] = ItsmStyle.builder().value(subjectScopes.joinToString(",") { it.name }).build()
-        val itsmValue = ItsmValue.builder()
-            .scheme("content_table")
-            .lable(
-                I18nUtil.getCodeLanMessage(BK_CREATE_PROJECT_APPROVAL)
-            )
-            .value(listOf(value))
-            .build()
-        return ItsmContentDTO.builder().formData(Arrays.asList(itsmValue)).schemes(scheme).build()
     }
 }
